# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Khalaf Olive Groves - Ancient Grove Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: assignedToUserId
          schema:
            type: integer
        - in: query
          name: groveId
          schema:
            type: integer
        - in: query
          name: treeId
          schema:
            type: integer
        - in: query
          name: taskType
          schema:
            type: string
        - in: query
          name: satelliteAlertId
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Cancel/delete task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # ─── Field Visits ─────────────────────────────────────────────────────────────
...
  /traps:
    get:
      operationId: listTraps
      tags: [traps]
      summary: List traps (optionally filtered by grove)
      parameters:
        - in: query
          name: groveId
          schema: { type: integer }
        - in: query
          name: kind
          schema: { type: string }
        - in: query
          name: activeOnly
          schema: { type: boolean }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Traps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trap"
    post:
      operationId: createTrap
      tags: [traps]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrapRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trap"
  /traps/{id}:
    get:
      operationId: getTrap
      tags: [traps]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Trap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trap"
    patch:
      operationId: updateTrap
      tags: [traps]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrapRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trap"
    delete:
      operationId: deleteTrap
      tags: [traps]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /trap-counts:
    get:
      operationId: listTrapCounts
      tags: [traps]
      summary: List trap counts
      parameters:
        - in: query
          name: trapId
          schema: { type: integer }
        - in: query
          name: groveId
          schema: { type: integer }
        - in: query
          name: fromDate
          schema: { type: string }
        - in: query
          name: toDate
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrapCount"
    post:
      operationId: createTrapCount
      tags: [traps]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrapCountRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrapCount"
  /trap-counts/{id}:
    get:
      operationId: getTrapCount
      tags: [traps]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Trap count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrapCount"
    patch:
      operationId: updateTrapCount
      tags: [traps]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrapCountRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrapCount"
    delete:
      operationId: deleteTrapCount
      tags: [traps]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Treatments ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        taskType:
          type: string
          enum: [tree_inspection, grove_inspection, terrace_check, pest_check, drought_check, pruning_check, harvest_check, understory_survey, lab_sampling, pressing_logistics, satellite_alert_review, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, assigned, in_progress, completed, cancelled]
        assignedToUserId:
          type: integer
          nullable: true
        assignedToUserName:
          type: string
          nullable: true
        groveId:
          type: integer
          nullable: true
        groveName:
          type: string
          nullable: true
        treeId:
          type: integer
          nullable: true
        treeCode:
          type: string
          nullable: true
        satelliteAlertId:
          type: integer
          nullable: true
        heritageRuleId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdByUserId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - taskType
        - priority
        - status
        - createdAt
        - updatedAt
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        taskType:
          type: string
          enum: [tree_inspection, grove_inspection, terrace_check, pest_check, drought_check, pruning_check, harvest_check, understory_survey, lab_sampling, pressing_logistics, satellite_alert_review, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedToUserId:
          type: integer
          nullable: true
        groveId:
          type: integer
          nullable: true
        treeId:
          type: integer
          nullable: true
        satelliteAlertId:
          type: integer
          nullable: true
        heritageRuleId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        createdByUserId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - title
        - taskType
        - priority
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, assigned, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedToUserId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true

    # ─── Field Visit ───────────────────────────────────────────────────────────
...
    Trap:
      type: object
      required: [id, groveId, code, kind, installedAt, createdAt]
      properties:
        id: { type: integer }
        groveId: { type: integer }
        groveName: { type: string, nullable: true }
        code: { type: string }
        kind:
          type: string
          enum: [pheromone, sticky_yellow, sticky_red, mcphail, light, food_bait, other]
        targetSpecies: { type: string, nullable: true }
        locationLat: { type: number, nullable: true }
        locationLon: { type: number, nullable: true }
        locationDescription: { type: string, nullable: true }
        installedAt: { type: string, format: date-time }
        retiredAt: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
        latestCount: { type: integer, nullable: true }
        latestCountDate: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateTrapRequest:
      type: object
      required: [groveId, code, kind]
      properties:
        groveId: { type: integer }
        code: { type: string }
        kind:
          type: string
          enum: [pheromone, sticky_yellow, sticky_red, mcphail, light, food_bait, other]
        targetSpecies: { type: string, nullable: true }
        locationLat: { type: number, nullable: true }
        locationLon: { type: number, nullable: true }
        locationDescription: { type: string, nullable: true }
        installedAt: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
    UpdateTrapRequest:
      type: object
      properties:
        code: { type: string, nullable: true }
        kind:
          type: string
          enum: [pheromone, sticky_yellow, sticky_red, mcphail, light, food_bait, other]
          nullable: true
        targetSpecies: { type: string, nullable: true }
        locationLat: { type: number, nullable: true }
        locationLon: { type: number, nullable: true }
        locationDescription: { type: string, nullable: true }
        retiredAt: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
    TrapCount:
      type: object
      required: [id, trapId, workerId, count, countDate, source, createdAt]
      properties:
        id: { type: integer }
        trapId: { type: integer }
        trapCode: { type: string, nullable: true }
        trapKind: { type: string, nullable: true }
        groveId: { type: integer, nullable: true }
        groveName: { type: string, nullable: true }
        workerId: { type: integer }
        workerName: { type: string, nullable: true }
        count: { type: integer }
        countDate: { type: string, format: date-time }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        source:
          type: string
          enum: [manual, photo_count, sensor]
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateTrapCountRequest:
      type: object
      required: [trapId, workerId, count]
      properties:
        trapId: { type: integer }
        workerId: { type: integer }
        count: { type: integer }
        countDate: { type: string, format: date-time, nullable: true }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        source:
          type: string
          enum: [manual, photo_count, sensor]
          nullable: true
        notes: { type: string, nullable: true }
    UpdateTrapCountRequest:
      type: object
      properties:
        count: { type: integer, nullable: true }
        countDate: { type: string, format: date-time, nullable: true }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        source:
          type: string
          enum: [manual, photo_count, sensor]
          nullable: true
        notes: { type: string, nullable: true }
...